chore: Move test files
